+++ ext/nokogiri/xml_xpath_context.c
@@ -186,29 +186,29 @@ noko_xml_xpath_context_register_variable(VALUE rb_context, VALUE name, VALUE val
 }
 
 
-/*
- *  convert an XPath object into a Ruby object of the appropriate type.
- *  returns Qundef if no conversion was possible.
- */
-static VALUE
-_noko_xml_xpath_context__xpath2ruby(xmlXPathObjectPtr c_xpath_object, xmlXPathContextPtr c_context)
-{
-  VALUE rb_retval;
+typedef struct {
+  xmlXPathObjectPtr xpath_object;
+  xmlXPathContextPtr context;
+  xmlChar *string;
+  int string_fallback;
+} xpath2ruby_args;
 
-  assert(c_context->doc);
-  assert(DOC_RUBY_OBJECT_TEST(c_context->doc));
+/* Convert an XPath object, transferring node set ownership to its Ruby wrapper. */
+static VALUE
+_noko_xml_xpath_context__xpath2ruby_convert(VALUE data)
+{
+  xpath2ruby_args *args = (xpath2ruby_args *)data;
+  xmlXPathObjectPtr c_xpath_object = args->xpath_object;
 
   switch (c_xpath_object->type) {
     case XPATH_STRING:
-      rb_retval = NOKOGIRI_STR_NEW2(c_xpath_object->stringval);
-      xmlFree(c_xpath_object->stringval);
-      return rb_retval;
+      return NOKOGIRI_STR_NEW2(c_xpath_object->stringval);
 
-    case XPATH_NODESET:
-      return noko_xml_node_set_wrap(
-               c_xpath_object->nodesetval,
-               DOC_RUBY_OBJECT(c_context->doc)
-             );
+    case XPATH_NODESET: {
+      xmlNodeSetPtr node_set = c_xpath_object->nodesetval;
+      c_xpath_object->nodesetval = NULL;
+      return noko_xml_node_set_wrap(node_set, DOC_RUBY_OBJECT(args->context->doc));
+    }
 
     case XPATH_NUMBER:
       return rb_float_new(c_xpath_object->floatval);
@@ -217,8 +217,39 @@ _noko_xml_xpath_context__xpath2ruby(xmlXPathObjectPtr c_xpath_object, xmlXPathCo
       return (c_xpath_object->boolval == 1) ? Qtrue : Qfalse;
 
     default:
-      return Qundef;
+      if (args->string_fallback) {
+        args->string = xmlXPathCastToString(c_xpath_object);
+        return NOKOGIRI_STR_NEW2(args->string);
+      }
+      return noko_xml_node_set_wrap(NULL, DOC_RUBY_OBJECT(args->context->doc));
   }
+}
+
+static VALUE
+_noko_xml_xpath_context__xpath2ruby_cleanup(VALUE data)
+{
+  xpath2ruby_args *args = (xpath2ruby_args *)data;
+  xmlFree(args->string);
+  xmlXPathFreeObject(args->xpath_object);
+  return Qnil;
+}
+
+static VALUE
+_noko_xml_xpath_context__xpath2ruby(xmlXPathObjectPtr c_xpath_object, xmlXPathContextPtr c_context,
+                                    int string_fallback)
+{
+  assert(c_context->doc);
+  assert(DOC_RUBY_OBJECT_TEST(c_context->doc));
+
+  xpath2ruby_args args = {
+    .xpath_object = c_xpath_object,
+    .context = c_context,
+    .string_fallback = string_fallback,
+  };
+  return rb_ensure(
+           _noko_xml_xpath_context__xpath2ruby_convert, (VALUE)&args,
+           _noko_xml_xpath_context__xpath2ruby_cleanup, (VALUE)&args
+         );
 }
 
 static xmlNodeSetPtr
@@ -259,11 +290,7 @@ Nokogiri_marshal_xpath_funcall_and_return_values(
 
   for (int j = argc - 1 ; j >= 0 ; --j) {
     c_xpath_object = valuePop(ctxt);
-    argv[j] = _noko_xml_xpath_context__xpath2ruby(c_xpath_object, ctxt->context);
-    if (argv[j] == Qundef) {
-      argv[j] = NOKOGIRI_STR_NEW2(xmlXPathCastToString(c_xpath_object));
-    }
-    xmlXPathFreeNodeSetList(c_xpath_object);
+    argv[j] = _noko_xml_xpath_context__xpath2ruby(c_xpath_object, ctxt->context, 1);
   }
 
   rb_retval = rb_funcall2(
@@ -313,44 +340,66 @@ Nokogiri_marshal_xpath_funcall_and_return_values(
 typedef struct {
   VALUE handler;
   VALUE retained_nodes;
+  int state;
+  xmlXPathParserContextPtr ctxt;
+  int argc;
+  const xmlChar *name;
+  const xmlChar *ns_uri;
 } xpath_handler_args;
+
+static VALUE
+_noko_xml_xpath_context__handler_invoker_protected(VALUE data)
+{
+  xpath_handler_args *args = (xpath_handler_args *)data;
+  Nokogiri_marshal_xpath_funcall_and_return_values(
+    args->ctxt, args->argc, args->handler, (const char *)args->ctxt->context->function,
+    args->retained_nodes
+  );
+  return Qnil;
+}
 
 static void
 _noko_xml_xpath_context__handler_invoker(xmlXPathParserContextPtr ctxt, int argc)
 {
-  xpath_handler_args *args = NULL;
-  const char *method_name = NULL ;
+  xpath_handler_args *args = ctxt->context->userData;
 
-  assert(ctxt);
-  assert(ctxt->context);
-  assert(ctxt->context->userData);
-  assert(ctxt->context->function);
+  if (!args->state) {
+    args->ctxt = ctxt;
+    args->argc = argc;
+    rb_protect(_noko_xml_xpath_context__handler_invoker_protected, (VALUE)args, &args->state);
+  }
+  if (args->state) {
+    /* Let libxml2 unwind predicates and restore its evaluation state before the Ruby jump. */
+    ctxt->error = XPATH_EXPR_ERROR;
+  }
+}
 
-  args = ctxt->context->userData;
-  method_name = (const char *)(ctxt->context->function);
+static VALUE
+_noko_xml_xpath_context_handler_lookup_protected(VALUE data)
+{
+  xpath_handler_args *args = (xpath_handler_args *)data;
+  if (rb_respond_to(args->handler, rb_intern((const char *)args->name))) {
+    if (args->ns_uri == NULL) {
+      NOKO_WARN_DEPRECATION("A custom XPath or CSS handler function named '%s' is being invoked without a namespace. Please update your query to reference this function as 'nokogiri:%s'. Invoking custom handler functions without a namespace is deprecated and will become an error in Nokogiri v1.17.0.",
+                            args->name, args->name); // TODO deprecated in v1.15.0, remove in v1.19.0
+    }
+    return Qtrue;
+  }
 
-  Nokogiri_marshal_xpath_funcall_and_return_values(
-    ctxt,
-    argc,
-    args->handler,
-    method_name,
-    args->retained_nodes
-  );
+  return Qfalse;
 }
 
 static xmlXPathFunction
 _noko_xml_xpath_context_handler_lookup(void *data, const xmlChar *c_name, const xmlChar *c_ns_uri)
 {
-  VALUE rb_handler = ((xpath_handler_args *)data)->handler;
-  if (rb_respond_to(rb_handler, rb_intern((const char *)c_name))) {
-    if (c_ns_uri == NULL) {
-      NOKO_WARN_DEPRECATION("A custom XPath or CSS handler function named '%s' is being invoked without a namespace. Please update your query to reference this function as 'nokogiri:%s'. Invoking custom handler functions without a namespace is deprecated and will become an error in Nokogiri v1.17.0.",
-                            c_name, c_name); // TODO deprecated in v1.15.0, remove in v1.19.0
-    }
-    return _noko_xml_xpath_context__handler_invoker;
+  xpath_handler_args *args = data;
+  VALUE result = Qfalse;
+  if (!args->state) {
+    args->name = c_name;
+    args->ns_uri = c_ns_uri;
+    result = rb_protect(_noko_xml_xpath_context_handler_lookup_protected, (VALUE)args, &args->state);
   }
-
-  return NULL;
+  return args->state || RTEST(result) ? _noko_xml_xpath_context__handler_invoker : NULL;
 }
 
 PRINTFLIKE_DECL(2, 3)
@@ -378,6 +427,13 @@ _noko_xml_xpath_context__generic_exception_pusher(void *data, const char *msg, .
   rb_ary_push(rb_errors, rb_exception);
 }
 
+static VALUE
+_noko_xml_xpath_context_evaluate_protected(VALUE data)
+{
+  xmlXPathEvalExpr((xmlXPathParserContextPtr)data);
+  return Qnil;
+}
+
 /*
  * call-seq:
  *   evaluate(search_path, handler = nil) → Object
@@ -395,8 +451,9 @@ noko_xml_xpath_context_evaluate(int argc, VALUE *argv, VALUE rb_context)
   VALUE rb_function_lookup_handler = Qnil;
   xmlChar *c_expression_str = NULL;
   VALUE rb_errors = rb_ary_new();
-  xmlXPathObjectPtr c_xpath_object;
+  xmlXPathObjectPtr c_xpath_object = NULL;
   VALUE rb_xpath_object = Qnil;
+  int state = 0;
   libxmlStructuredErrorHandlerState handler_state;
   xmlGenericErrorFunc previous_generic_handler = xmlGenericError;
   void *previous_generic_context = xmlGenericErrorContext;
@@ -432,26 +489,43 @@ noko_xml_xpath_context_evaluate(int argc, VALUE *argv, VALUE rb_context)
   noko__structured_error_func_save_and_set(&handler_state, (void *)rb_errors, noko__error_array_pusher);
   xmlSetGenericErrorFunc((void *)rb_errors, _noko_xml_xpath_context__generic_exception_pusher);
 
-  c_xpath_object = xmlXPathEvalExpression(c_expression_str, c_context);
+  /* Own the parser context so a Ruby callback cannot skip its cleanup. */
+  xmlResetError(&c_context->lastError);
+  xmlXPathParserContextPtr c_parser = xmlXPathNewParserContext(c_expression_str, c_context);
+  if (c_parser) {
+    rb_protect(_noko_xml_xpath_context_evaluate_protected, (VALUE)c_parser, &state);
+    if (handler_args.state) {
+      state = handler_args.state;
+    }
+    if (!state && c_parser->error == XPATH_EXPRESSION_OK) {
+      if (c_parser->valueNr == 1) {
+        c_xpath_object = valuePop(c_parser);
+      } else {
+        xmlXPathErr(c_parser, XPATH_STACK_ERROR);
+      }
+    }
+    xmlXPathFreeParserContext(c_parser);
+  }
 
   noko__structured_error_func_restore(&handler_state);
   xmlSetGenericErrorFunc(previous_generic_context, previous_generic_handler);
 
   xmlXPathRegisterFuncLookup(c_context, previous_lookup, previous_lookup_data);
   c_context->userData = previous_user_data;
+  RB_GC_GUARD(rb_expression);
+  RB_GC_GUARD(rb_function_lookup_handler);
+  RB_GC_GUARD(rb_retained_nodes);
+  if (state) {
+    rb_jump_tag(state);
+  }
 
   if (c_xpath_object == NULL) {
     rb_exc_raise(rb_ary_entry(rb_errors, 0));
   }
 
-  rb_xpath_object = _noko_xml_xpath_context__xpath2ruby(c_xpath_object, c_context);
-  if (rb_xpath_object == Qundef) {
-    rb_xpath_object = noko_xml_node_set_wrap(NULL, DOC_RUBY_OBJECT(c_context->doc));
-  }
-
-  xmlXPathFreeNodeSetList(c_xpath_object);
-  RB_GC_GUARD(rb_function_lookup_handler);
+  rb_xpath_object = _noko_xml_xpath_context__xpath2ruby(c_xpath_object, c_context, 0);
   RB_GC_GUARD(rb_retained_nodes);
+  RB_GC_GUARD(rb_context);
 
   return rb_xpath_object;
 }